Add Community Configuration and i18n messages
This patch allows the extension to be enabled / disabled from Community
Configuration. Furthermore, it adds utilization of onBeforePageDisplay
hook in order to make clientPrefs work when the respective preferences
will be added to Vector skin and MobileFrontend extension (in order to
support Minerva). Lastly, all the needed i18n messages are added, so the
respective Community Configuration options, and user preferences can be
rendered.

Bug: T408435
Change-Id: I4b745362f5460616976895d9dc7f2086a8fb79c9